Update categories.php
This is ACL related. Whenever a user does not have superuser privileges, but has red write access to com_content, it is not allowed to change the ordering of the categories, because of insufficient write access. To get this working I found out that a.o. it is required to check at this level for 'com_content' authorisation. Key question is why does JRequest::getCmd('extension') does not provide this.